+++ registry/src/endpoints/mod.rs
@@ -1,4 +1,7 @@
 pub mod deprecate_version;
+pub mod package_archive;
+pub mod package_doc;
+pub mod package_readme;
 pub mod package_version;
 pub mod package_versions;
 pub mod publish_version;