+++ registry/src/endpoints/package_version.rs
@@ -2,13 +2,14 @@ use actix_web::{http::header::ACCEPT, web, HttpRequest, HttpResponse};
 use serde::Deserialize;
 
 use crate::{
+	endpoints::package_doc::find_package_doc,
 	error::RegistryError,
 	package::{read_package, PackageResponse},
-	request_path::{AnyOrSpecificTarget, LatestOrSpecificVersion},
+	request_path::{resolve_version_and_target, AnyOrSpecificTarget, LatestOrSpecificVersion},
 	storage::StorageImpl,
 	AppState,
 };
-use pesde::{names::PackageName, source::pesde::DocEntryKind};
+use pesde::names::PackageName;
 
 #[derive(Debug, Deserialize)]
 pub struct Query {
@@ -27,68 +28,38 @@ pub async fn get_package_version(
 		return Ok(HttpResponse::NotFound().finish());
 	};
 
-	let Some((v_id, entry)) = ({
-		let version = match version {
-			LatestOrSpecificVersion::Latest => match file.entries.keys().map(|k| k.version()).max()
-			{
-				Some(latest) => latest.clone(),
-				None => return Ok(HttpResponse::NotFound().finish()),
-			},
-			LatestOrSpecificVersion::Specific(version) => version,
-		};
-
-		let mut versions = file
-			.entries
-			.iter()
-			.filter(|(v_id, _)| *v_id.version() == version);
-
-		match target {
-			AnyOrSpecificTarget::Any => versions.min_by_key(|(v_id, _)| *v_id.target()),
-			AnyOrSpecificTarget::Specific(kind) => {
-				versions.find(|(_, entry)| entry.target.kind() == kind)
-			}
-		}
-	}) else {
+	let Some(v_id) = resolve_version_and_target(&file, version, target) else {
 		return Ok(HttpResponse::NotFound().finish());
 	};
 
-	if let Some(doc_name) = request_query.doc.as_deref() {
-		let hash = 'finder: {
-			let mut queue = entry.docs.iter().map(|doc| &doc.kind).collect::<Vec<_>>();
-			while let Some(doc) = queue.pop() {
-				match doc {
-					DocEntryKind::Page { name, hash } if name == doc_name => {
-						break 'finder hash.clone()
-					}
-					DocEntryKind::Category { items, .. } => {
-						queue.extend(items.iter().map(|item| &item.kind))
-					}
-					_ => continue,
-				};
-			}
+	// TODO: this is deprecated, since the introduction of the specific endpoints for readme, doc and archive.
+	//	     remove this when we drop 0.5 support.
+	{
+		if let Some(doc_name) = request_query.doc.as_deref() {
+			let Some(hash) = find_package_doc(&file, v_id, doc_name) else {
+				return Ok(HttpResponse::NotFound().finish());
+			};
 
-			return Ok(HttpResponse::NotFound().finish());
-		};
+			return app_state.storage.get_doc(hash).await;
+		}
 
-		return app_state.storage.get_doc(&hash).await;
-	}
+		let accept = request
+			.headers()
+			.get(ACCEPT)
+			.and_then(|accept| accept.to_str().ok())
+			.and_then(|accept| match accept.to_lowercase().as_str() {
+				"text/plain" => Some(true),
+				"application/octet-stream" => Some(false),
+				_ => None,
+			});
 
-	let accept = request
-		.headers()
-		.get(ACCEPT)
-		.and_then(|accept| accept.to_str().ok())
-		.and_then(|accept| match accept.to_lowercase().as_str() {
-			"text/plain" => Some(true),
-			"application/octet-stream" => Some(false),
-			_ => None,
-		});
-
-	if let Some(readme) = accept {
-		return if readme {
-			app_state.storage.get_readme(&name, v_id).await
-		} else {
-			app_state.storage.get_package(&name, v_id).await
-		};
+		if let Some(readme) = accept {
+			return if readme {
+				app_state.storage.get_readme(&name, v_id).await
+			} else {
+				app_state.storage.get_package(&name, v_id).await
+			};
+		}
 	}
 
 	Ok(HttpResponse::Ok().json(PackageResponse::new(&name, v_id, &file)))