+++ registry/src/request_path.rs
@@ -1,4 +1,7 @@
-use pesde::manifest::target::TargetKind;
+use pesde::{
+	manifest::target::TargetKind,
+	source::{ids::VersionId, pesde::IndexFile},
+};
 use semver::Version;
 use serde::{Deserialize, Deserializer};
 
@@ -44,6 +47,33 @@ impl<'de> Deserialize<'de> for AnyOrSpecificTarget {
 			.map(AnyOrSpecificTarget::Specific)
 			.map_err(serde::de::Error::custom)
 	}
+}
+
+pub fn resolve_version_and_target(
+	file: &IndexFile,
+	version: LatestOrSpecificVersion,
+	target: AnyOrSpecificTarget,
+) -> Option<&VersionId> {
+	let version = match version {
+		LatestOrSpecificVersion::Latest => match file.entries.keys().map(|k| k.version()).max() {
+			Some(latest) => latest.clone(),
+			None => return None,
+		},
+		LatestOrSpecificVersion::Specific(version) => version,
+	};
+
+	let mut versions = file
+		.entries
+		.iter()
+		.filter(|(v_id, _)| *v_id.version() == version);
+
+	match target {
+		AnyOrSpecificTarget::Any => versions.min_by_key(|(v_id, _)| *v_id.target()),
+		AnyOrSpecificTarget::Specific(kind) => {
+			versions.find(|(_, entry)| entry.target.kind() == kind)
+		}
+	}
+	.map(|(v_id, _)| v_id)
 }
 
 #[derive(Debug)]