+++ registry/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support deprecating and yanking packages by @daimond113
 - Add yanking & deprecating to registry by @daimond113
 - Log more information about configured auth & storage by @daimond113
+- Add individual endpoints for package data over using `Accept` header conditional returns by @daimond113
 
 ### Performance
 - Switch to using a `RwLock` over a `Mutex` to store repository data by @daimond113